requirements.txt
requests~=2.28.2
certifi~=2022.12.7
geopy~=2.3.0
pandas~=2.0.1
apache-airflow~=2.6.0